requirements.txt
beautifulsoup4
requests
pyrootutils
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
caldav